backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
python-dateutil==2.8.2
